+++ test/git_hub_actions/config_test.exs
@@ -1,5 +1,5 @@
 defmodule GitHubActions.ConfigTest do
-  use ExUnit.Case
+  use ExUnit.Case, async: true
 
   import Prove