+++ test/git_hub_actions/project_test.exs
@@ -1,5 +1,5 @@
 defmodule GitHubActions.ProjectTest do
-  use ExUnit.Case
+  use ExUnit.Case, async: true
 
   alias GitHubActions.Project