+++ test/git_hub_actions/sigils_test.exs
@@ -1,7 +1,5 @@
-# credo:disable-for-this-file
-# Reactivate when credo version > 1.6.2 is released.
 defmodule GitHubActions.SigilsTest do
-  use ExUnit.Case
+  use ExUnit.Case, async: true
 
   import GitHubActions.Sigils