+++ test/git_hub_actions/version_test.exs
@@ -1,5 +1,5 @@
 defmodule GitHubActions.VersionTest do
-  use ExUnit.Case
+  use ExUnit.Case, async: true
 
   import Prove