+++ test/git_hub_actions/versions_test.exs
@@ -1,5 +1,5 @@
 defmodule GitHubActions.VersionsTest do
-  use ExUnit.Case
+  use ExUnit.Case, async: true
 
   import Prove
 
@@ -19,7 +19,7 @@ defmodule GitHubActions.VersionsTest do
 
   test "from_config/0" do
     assert List.last(Versions.from_config()) ==
-             [otp: ["27.0/2"], elixir: ["1.17.0/3", "1.18.0/2"]]
+             [otp: ["27.0/3"], elixir: ["1.17.0/3", "1.18.0/3"]]
   end
 
   describe "get/2" do
@@ -135,10 +135,11 @@ defmodule GitHubActions.VersionsTest do
                "1.17.1",
                "1.17.2",
                "1.17.3",
-               # v1.18.0/2
+               # v1.18.0/3
                "1.18.0",
                "1.18.1",
-               "1.18.2"
+               "1.18.2",
+               "1.18.3"
              ]
     end
 
@@ -188,7 +189,8 @@ defmodule GitHubActions.VersionsTest do
                  "26.2",
                  "27.0",
                  "27.1",
-                 "27.2"
+                 "27.2",
+                 "27.3"
                ]
     end
   end
@@ -294,7 +296,8 @@ defmodule GitHubActions.VersionsTest do
                %Version{major: 1, minor: 17, patch: 3},
                %Version{major: 1, minor: 18, patch: 0},
                %Version{major: 1, minor: 18, patch: 1},
-               %Version{major: 1, minor: 18, patch: 2}
+               %Version{major: 1, minor: 18, patch: 2},
+               %Version{major: 1, minor: 18, patch: 3}
              ]
     end
 
@@ -332,7 +335,7 @@ defmodule GitHubActions.VersionsTest do
                %Version{major: 1, minor: 15, patch: 8},
                %Version{major: 1, minor: 16, patch: 3},
                %Version{major: 1, minor: 17, patch: 3},
-               %Version{major: 1, minor: 18, patch: 2}
+               %Version{major: 1, minor: 18, patch: 3}
              ]
     end
   end
@@ -350,7 +353,7 @@ defmodule GitHubActions.VersionsTest do
                %Version{major: 24, minor: 3},
                %Version{major: 25, minor: 3},
                %Version{major: 26, minor: 2},
-               %Version{major: 27, minor: 2}
+               %Version{major: 27, minor: 3}
              ]
     end
   end
@@ -463,7 +466,7 @@ defmodule GitHubActions.VersionsTest do
                ],
                [
                  otp: %Version{major: 22, minor: 3},
-                 elixir: %Version{major: 1, minor: 18, patch: 2}
+                 elixir: %Version{major: 1, minor: 18, patch: 3}
                ],
                [
                  otp: %Version{major: 23, minor: 3},
@@ -479,7 +482,7 @@ defmodule GitHubActions.VersionsTest do
                ],
                [
                  otp: %Version{major: 23, minor: 3},
-                 elixir: %Version{major: 1, minor: 18, patch: 2}
+                 elixir: %Version{major: 1, minor: 18, patch: 3}
                ],
                [
                  otp: %Version{major: 24, minor: 3},
@@ -487,7 +490,7 @@ defmodule GitHubActions.VersionsTest do
                ],
                [
                  otp: %Version{major: 24, minor: 3},
-                 elixir: %Version{major: 1, minor: 18, patch: 2}
+                 elixir: %Version{major: 1, minor: 18, patch: 3}
                ],
                [
                  otp: %Version{major: 25, minor: 3},
@@ -502,23 +505,23 @@ defmodule GitHubActions.VersionsTest do
                  elixir: %Version{major: 1, minor: 13, patch: 4}
                ],
                [
-                 otp: %Version{major: 27, minor: 2},
+                 otp: %Version{major: 27, minor: 3},
                  elixir: %Version{major: 1, minor: 12, patch: 3}
                ],
                [
-                 otp: %Version{major: 27, minor: 2},
+                 otp: %Version{major: 27, minor: 3},
                  elixir: %Version{major: 1, minor: 13, patch: 4}
                ],
                [
-                 otp: %Version{major: 27, minor: 2},
+                 otp: %Version{major: 27, minor: 3},
                  elixir: %Version{major: 1, minor: 14, patch: 5}
                ],
                [
-                 otp: %Version{major: 27, minor: 2},
+                 otp: %Version{major: 27, minor: 3},
                  elixir: %Version{major: 1, minor: 15, patch: 8}
                ],
                [
-                 otp: %Version{major: 27, minor: 2},
+                 otp: %Version{major: 27, minor: 3},
                  elixir: %Version{major: 1, minor: 16, patch: 3}
                ]
              ]